Add PATCH /todos/:id route

+++ server/server.js
@@ -1,5 +1,6 @@
-var express = require('express');
-var bodyParser = require('body-parser');
+const _ = require('lodash');
+const express = require('express');
+const bodyParser = require('body-parser');
 const { ObjectID } = require('mongodb');
 
 var { mongoose } = require('./db/mongoose');
@@ -59,15 +60,43 @@ app.get('/todos/:id', (req, res) =>{
 
 app.delete('/todos/:id', (req, res) => {
   var id = req.params.id;
+  if (!ObjectID.isValid(id)) {
+    return res.status(404).send({});
+  }
+
   Todo.findOneAndRemove({_id: id}).then((removedDoc) => {
     if (!removedDoc) {
         return res.status(404).send({});
     }
-    res.send(removedDoc);
+    res.send({removedDoc});
   }, (e) => res.status(400).send(e));
 
 });
 
+app.patch('/todos/:id', (req, res) => {
+  var id = req.params.id;
+  var body = _.pick(req.body, ['text', 'completed']);
+
+  if (!ObjectID.isValid(id)) {
+    return res.status(404).send({});
+  }
+
+  if (_.isBoolean(body.completed) && body.completed) {
+    body.completedAt = new Date().getTime();
+  } else {
+    body.completed = false;
+    body.completedAt = null;
+  }
+
+  Todo.findByIdAndUpdate(id, {$set: body}, {new: true}).then((todo) => {
+    if (!todo) {
+      return res.status(404).send();
+    }
+
+    return res.send({todo});
+  }).catch((e)=> res.status(400).send(e));
+
+});
 
 app.listen(port, () => {
   console.log(`Server is lestening at port ${port}...`);
